google-cloud/functions_storage_uploaded/requirements.txt
functions-framework
pillow
google-cloud-storage
google-cloud-vision
requests